requirements.txt
# ==================== Monolith Lite Requirements ====================
# All dependencies for Agent + 5 internal MCP modules

# ==================== Web Framework ====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# ==================== HTTP & Network ====================
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0

# ==================== AI/ML & NLP ====================
torch==2.1.1
transformers==4.35.2
huggingface_hub==0.19.4
numpy==1.24.3
scipy==1.11.4

# ==================== LangGraph & Chain ====================
langgraph==0.0.15
langchain==0.0.340
langchain-core==0.0.14

# ==================== Graph Analysis ====================
networkx==3.2.1

# ==================== Code Analysis & AST ====================
tree-sitter==0.20.1
tree-sitter-python==0.20.4

# ==================== Optional (for local model inference) ====================
# torch and transformers above enable local model inference
# Uncomment if using HuggingFace API instead of local models:
# huggingface_hub>=0.19.0
